backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Data validation and settings
pydantic==2.10.3
pydantic-settings==2.6.1

# Environment variables
python-dotenv==1.0.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# LLM Provider SDKs
openai==1.59.6
anthropic==0.42.0
google-genai==1.7.0
requests==2.32.3

# Code quality
black==24.10.0
ruff==0.8.2
mypy==1.13.0
